pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "cc-analysis"
version = "0.1.0"
description = "Rule-based analysis engine for the China Compass pipeline"
authors = ["China Compass Team"]
readme = "README.md"
packages = [{include = "analysis", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0"
click = "^8.1"
jsonschema = "^4.20"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.5"

[tool.poetry.scripts]
analysis = "analysis.cli:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
